pyproject.toml


[project]

requires-python = ">=3.10"

name = "transpolymer-pretrained-polyagent"
version = "1.0.0"
description = "TransPolymer: A Transformer-based language model for polymer property predictions"
readme = "README.md"
license = {text = "Apache License 2.0"}



dependencies = [
    "torch>=1.12.0",
    "transformers>=4.20.1",
    "pyyaml>=6.0",
    "fairscale>=0.4.6",
    "rdkit>=2022.3.4",
    "scikit-learn>=0.24.2",
    "pandas>=1.0.0",
    "numpy>=1.19.0",
    "joblib>=1.0.0",
    "regex>=2022.0.0",
    "tensorboard>=2.20.0",
    "ipykernel>=7.1.0",
    "mcp>=1.21.1",
    "torchmetrics>=1.8.2",
]

[project.optional-dependencies]
visualization = [
    "tensorboard>=2.9.1",
    "seaborn>=0.11.2",
    "matplotlib>=3.3.0",
    "opentsne>=0.6.2",
]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]
mcp = [
    "mcp>=0.3.0",
]

[project.urls]
Homepage = "https://github.com/ChangwenXu98/TransPolymer"
Repository = "https://github.com/ChangwenXu98/TransPolymer.git"
Documentation = "https://github.com/ChangwenXu98/TransPolymer"
"Bug Tracker" = "https://github.com/ChangwenXu98/TransPolymer/issues"

[project.scripts]
transpolymer = "transpolymer.core.predict:main"
transpolymer-predict = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
[tool.setuptools.packages.find]
where = ["."]
include = ["transpolymer_pretrained*"]

[tool.setuptools.package-data]
transpolymer = [
    "configs/*.yaml",
    "ckpt/*",
    "data/*",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false


